+++ code/jump.hpp
@@ -4,7 +4,7 @@
 /* gsc functions */
 #include "gsc.hpp"
 
-double Jump_ReduceFriction(playerState_t *ps);
+float Jump_ReduceFriction(playerState_t *ps);
 void Jump_ClampVelocity(playerState_t* ps, vec3_t vec);
 qboolean Jump_IsPlayerAboveMax(playerState_t* ps);
 qboolean Jump_GetStepHeight(playerState_t* ps, const vec3_t vec1, float* val2);
